feat(dynamic-menu): exported footer components
