+++ packages/falcon-ecommerce-uikit/src/Footer/index.ts
@@ -1,3 +1,6 @@
+export * from './FooterQuery';
 export * from './Footer';
 export * from './Newsletter';
-export * from './FooterQuery';
+export * from './FooterSections';
+export * from './Copyright';
+export * from './LanguageSwitcher';
